Move DPU setup to conftest
